Add shims for ct_run and typer
Related to #38.

#38 adds some shims which have since been added - that pull request can be closed.

This pull request adds shims for ct_run.exe and typer.exe. The RabbitMQ team has been updating their test suites to run on Windows, and `ct_run.exe` needs to be in the `PATH`. If Erlang is installed via Chocolatey, it is not in the `PATH` without these shims.